+++ modules/hanna-react/src/IframeBlock.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import getBemClass from '@hugsmidjan/react/utils/getBemClass';
-import IframeResizer from 'iframe-resizer-react';
+import IframeResizer, { ResizerOptions } from 'iframe-resizer-react';
 
 export type IframeBlockProps = {
   src: string;
@@ -9,13 +9,16 @@ export type IframeBlockProps = {
   align?: 'right';
 } & (
   | {
-      /** Default: 'auto' ... which initializes "iframe-resizer" script */
+      /** Default: `'auto'` ... which initializes "iframe-resizer" script */
       height?: 'auto';
       scrolling?: never;
+      /** Default: `false` ... Set to `true` for same-site only, or provide array of allowed domain-names */
+      checkOrigin?: ResizerOptions['checkOrigin'];
     }
   | {
       height: number;
       scrolling?: boolean | 'no' | 'yes';
+      checkOrigin?: never;
     }
 );
 
@@ -29,7 +32,15 @@ export type IframeBlockProps = {
  */
 
 const IframeBlock = (props: IframeBlockProps) => {
-  const { src, framed, compact, scrolling, height = 'auto', align } = props;
+  const {
+    src,
+    framed,
+    compact,
+    scrolling,
+    height = 'auto',
+    align,
+    checkOrigin = false,
+  } = props;
 
   const className = getBemClass('IframeBlock', [
     framed && 'framed',
@@ -38,7 +49,7 @@ const IframeBlock = (props: IframeBlockProps) => {
   ]);
 
   return height === 'auto' ? (
-    <IframeResizer className={className} src={src} />
+    <IframeResizer className={className} src={src} checkOrigin={checkOrigin} />
   ) : (
     <iframe
       className={className}