+++ modules/hanna-react/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- fix: Add missing prop `checkOrigin` on `IframeBlock` (default: `false`)
 - fix: Suppot "uncontrolled" use of `CheckboxGroup`, `RadioGroup`, et al.
 
 ## 0.10.59